+++ vrtool/decision_making/measures/standard_measures/revetment_measure/revetment_measure.py
@@ -389,7 +389,9 @@ class RevetmentMeasure(MeasureProtocol):
         section_reliability.failure_mechanisms = _failure_mechanism_collection
         # TODO (VRTOOL-187)
         # Is this really required for revetments? Should we not better have our own RevetmentSectionReliability?
-        section_reliability.calculate_section_reliability()
+        section_reliability.calculate_section_reliability(
+            dike_section.get_cross_sectional_properties()
+        )
         return section_reliability, revetment_measure_results[0].cost
 
     def _get_mechanism_reliabilty_for_beta_transition(